refactor(card): split index.js into importable feature chunks
- hover.js: cursor-tracked hov(track|drift|tilt), standalone
- carousel.js: loop, autoplay, pause-on-slide-leave, per-slide <ui-play>
- video.js: embed facades, media commands, vid() tools, solo play, tracking
- shared.js: primitives shared by carousel.js/video.js
- index.js: thin all-in-one entry composing the chunks (same API as before)

Each chunk self-initializes on idle and is idempotent, so pages load only
what they use: media.carousel.html -> carousel.js, media.hover.html ->
hover.js, media.video.html -> video.js.

Add build.js (esbuild): bundles+minifies each entry to *.min.js and prints
a source/min/gzip/brotli size table. Brotli over the wire: index 4.0 kB,
carousel 1.9 kB, hover 0.5 kB, video 2.3 kB.
