+++ ui/card/media.furniture.html
@@ -271,6 +271,5 @@
 	</lay-out>
 	</section>
 
-	<script type="module" src="index.js"></script>
 </body>
 </html>
