+++ ui/card/media.hover.html
@@ -96,7 +96,7 @@
 
 	<section>
 	<h2>Cursor-driven — needs JS</h2>
-	<p>The only effects that need JavaScript: <code>index.js</code> sets <code>--ui-media-mx</code> / <code>--ui-media-my</code> (−1..1) on the frame from a <code>pointermove</code> handler; the CSS reads them.</p>
+	<p>The only effects that need JavaScript: <code>hover.js</code> sets <code>--ui-media-mx</code> / <code>--ui-media-my</code> (−1..1) on the frame from a <code>pointermove</code> handler; the CSS reads them.</p>
 	<lay-out md="columns(2)" lg="columns(3)">
 		<ui-card variant="col" tabindex="0" media="asr(16/9) hov(track)">
 			<cq-box>
@@ -119,6 +119,6 @@
 	</lay-out>
 	</section>
 
-	<script type="module" src="index.js"></script>
+	<script type="module" src="hover.js"></script>
 </body>
 </html>
