+++ ui/card/media.html
@@ -252,6 +252,6 @@
 	<p>Inspect the <code>&lt;img&gt;</code> in DevTools: each shows a <code>srcset</code> of <code>/cdn-cgi/image/format=avif,quality=80,fit=cover,width=…,height=…/assets/images/… NNNw</code> candidates (they 404 locally — that's expected; the URL strings are what matter). Remove <code>cdn="on"</code> and they stay as plain local images.</p> -->
 
 	<!-- cursor-tracked hover for hov(track)/hov(drift) — idle until hovered -->
-	<script type="module" src="index.js"></script>
+	<script type="module" src="hover.js"></script>
 </body>
 </html>
